requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
asttokens==3.0.1
async-lru==2.0.5
attrs==25.4.0
babel==2.17.0
backoff==2.2.1
bcrypt==5.0.0
beautifulsoup4==4.14.3
bleach==6.3.0
build==1.3.0
cachetools==6.2.2
certifi==2025.11.12
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.4
chromadb==1.3.6
click==8.3.1
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.3
contourpy==1.3.3
cycler==0.12.1
debugpy==1.8.18
decorator==5.2.1
defusedxml==0.7.1
distro==1.9.0
durationpy==0.10
executing==2.2.1
fastapi==0.124.2
fastjsonschema==2.21.2
filelock==3.20.0
flatbuffers==25.9.23
fonttools==4.61.0
fqdn==1.5.1
fsspec==2025.12.0
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.43.0
google-auth-httplib2==0.2.1
google-generativeai==0.8.5
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.71.2
gurobipy==13.0.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httptools==0.7.1
httpx==0.28.1
huggingface-hub==0.36.0
humanfriendly==10.0
idna==3.11
importlib_metadata==8.7.0
importlib_resources==6.5.2
ipykernel==7.1.0
ipython==9.8.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.8
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.2
json5==0.12.1
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.7.0
jupyter_core==5.9.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.3
jupyterlab==4.5.0
jupyterlab_pygments==0.3.0
jupyterlab_server==2.28.0
jupyterlab_widgets==3.0.16
kiwisolver==1.4.9
kubernetes==34.1.0
lark==1.3.1
llvmlite==0.46.0
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
matplotlib-inline==0.2.1
mdurl==0.1.2
mistune==3.1.4
mmh3==5.2.0
mpmath==1.3.0
narwhals==2.13.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.6.1
notebook==7.5.0
notebook_shim==0.2.4
numba==0.63.1
numpy==2.3.5
oauthlib==3.3.1
onnxruntime==1.23.2
opentelemetry-api==1.39.0
opentelemetry-exporter-otlp-proto-common==1.39.0
opentelemetry-exporter-otlp-proto-grpc==1.39.0
opentelemetry-proto==1.39.0
opentelemetry-sdk==1.39.0
opentelemetry-semantic-conventions==0.60b0
orbit @ file:///C:/Users/samca/Documents/Competitions/Quantum-RAG-Integrated/orbit/orbit-0.2.0-py3-none-any.whl
orjson==3.11.5
overrides==7.7.0
packaging==25.0
pandas==2.3.3
pandocfilters==1.5.1
parso==0.8.5
pillow==12.0.0
platformdirs==4.5.1
plotly==6.5.0
posthog==5.4.0
prometheus_client==0.23.1
prompt_toolkit==3.0.52
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.1.3
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.3
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pynndescent==0.5.13
pyparsing==3.2.5
PyPDF2==3.0.1
PyPika==0.48.9
pyproject_hooks==1.2.0
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.1
python-json-logger==4.0.0
pytz==2025.2
pywinpty==3.0.2
PyYAML==6.0.3
pyzmq==27.1.0
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==14.2.0
rpds-py==0.30.0
rsa==4.9.1
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.16.3
seaborn==0.13.2
Send2Trash==1.8.3
sentence-transformers==5.1.2
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
soupsieve==2.8
stack-data==0.6.3
starlette==0.50.0
sympy==1.14.0
tenacity==9.1.2
terminado==0.18.1
threadpoolctl==3.6.0
tinycss2==1.4.0
tokenizers==0.22.1
torch==2.9.1
tornado==6.5.3
tqdm==4.67.1
traitlets==5.14.3
transformers==4.57.3
typer==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
umap-learn==0.5.9.post2
uri-template==1.3.0
uritemplate==4.2.0
urllib3==2.3.0
uvicorn==0.38.0
watchfiles==1.1.1
wcwidth==0.2.14
webcolors==25.10.0
webencodings==0.5.1
websocket-client==1.9.0
websockets==15.0.1
widgetsnbextension==4.0.15
Wikipedia-API==0.8.1
zipp==3.23.0
